requirements.txt
test_basic.py# API ve HTTP işlemleri
aiohttp==3.9.1
requests==2.31.0

# Web scraping ve parsing
beautifulsoup4==4.12.2
html5lib==1.1
lxml==4.9.3

# Veri işleme ve analiz
python-dateutil==2.8.2
chardet==5.2.0

# Güvenlik ve kimlik doğrulama
fake-useragent==1.4.0
certifi==2023.11.17

# DNS ve ağ araçları
dnspython==2.4.2
python-whois==0.8.0

# Shodan API
shodan==1.30.1

# Dosya işleme
python-dotenv==1.0.0
aiofiles==23.2.1

# Yardımcı araçlar
tqdm==4.66.1
colorama==0.4.6